+++ tests/cli_commands/test_compile.py
@@ -1,6 +1,7 @@
 import json
 import os
 import pathlib
+import shutil
 import tempfile
 import unittest
 from typing import List
@@ -17,10 +18,6 @@ goldens_dir_path = pathlib.Path(__file__).parent.parent.joinpath("goldens")
 
 
 class CompileCommandTestCase(unittest.TestCase):
-    @staticmethod
-    def _datahub_content(ingest_endpoint):
-        return {"sink": {"config": {"server": f"https://{ingest_endpoint}:8080"}}}
-
     @staticmethod
     def _k8s_content(repository_url: str, tag: str):
         return {
@@ -196,3 +193,54 @@ class CompileCommandTestCase(unittest.TestCase):
         ):
             with self.assertRaises(DataPipelinesError):
                 compile_project("base", True)
+
+    @patch("data_pipelines_cli.data_structures.git_revision_hash")
+    def test_datahub_variables(self, mock_git_revision_hash):
+        commit_sha = "aaa9876aaa"
+        mock_git_revision_hash.return_value = commit_sha
+
+        datahub_content = {
+            "datahub_path": "{{ var('datahub_path') }}",
+            "heathers": {"api-key": "{{ env_var('SECRET_KEY') }}"},
+        }
+
+        runner = CliRunner()
+        with tempfile.TemporaryDirectory() as tmp_dir, patch(
+            "data_pipelines_cli.cli_commands.compile.BUILD_DIR", pathlib.Path(tmp_dir)
+        ), patch(
+            "data_pipelines_cli.config_generation.BUILD_DIR", pathlib.Path(tmp_dir)
+        ), patch(
+            "data_pipelines_cli.cli_constants.BUILD_DIR", pathlib.Path(tmp_dir)
+        ), patch(
+            "data_pipelines_cli.dbt_utils.BUILD_DIR", pathlib.Path(tmp_dir)
+        ), patch(
+            "data_pipelines_cli.dbt_utils.subprocess_run", self._mock_run
+        ), patch.dict(
+            "os.environ", SECRET_KEY="very_secret_keeeeeeeeeeeeeeeeeey.abcd"
+        ), tempfile.TemporaryDirectory() as tmp_dir_2, patch(
+            "pathlib.Path.cwd", lambda: pathlib.Path(tmp_dir_2)
+        ):
+            for dir in ["config", "dag", "target"]:
+                shutil.copytree(
+                    goldens_dir_path.joinpath(dir),
+                    pathlib.Path(tmp_dir_2).joinpath(dir),
+                )
+            with open(
+                pathlib.Path(tmp_dir_2).joinpath("config", "base", "datahub.yml"), "w"
+            ) as datahub_file:
+                yaml.dump(datahub_content, datahub_file)
+
+            result = runner.invoke(_cli, ["compile", "--env", "datahub"])
+            self.assertEqual(0, result.exit_code, msg=result.exception)
+
+            tmp_dir_path = pathlib.Path(tmp_dir)
+
+            expected_dict = {
+                "datahub_path": "http://example.com/datahub/RaNdOmTe$T__PAHT",
+                "heathers": {"api-key": "very_secret_keeeeeeeeeeeeeeeeeey.abcd"},
+            }
+
+            with open(
+                tmp_dir_path.joinpath("dag", "config", "base", "datahub.yml"), "r"
+            ) as tmp_datahub:
+                self.assertDictEqual(expected_dict, yaml.safe_load(tmp_datahub))